requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
PyMySQL
python-dotenv
Flask-Migrate
